pyproject.toml
[tool.poetry]
name = "edmundbotadder"
version = "0.1.0"
description = "A bot to convert episodes of Blackadder to Discord channels"
authors = ["Daniel Milnes <thebeanogamer@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
edmundbotadder = "edmundbotadder:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
